+++ sdk/servicebus/azure-servicebus/azure/servicebus/_servicebus_client.py
@@ -11,7 +11,7 @@ from ._servicebus_sender import ServiceBusSender
 from ._servicebus_receiver import ServiceBusReceiver
 from ._servicebus_session_receiver import ServiceBusSessionReceiver
 from ._common._configuration import Configuration
-from ._common.utils import create_authentication, generate_dead_letter_entity_name
+from ._common.utils import create_authentication
 
 if TYPE_CHECKING:
     from azure.core.credentials import TokenCredential
@@ -85,39 +85,6 @@ class ServiceBusClient(object):
             debug=self._config.logging_enable
         )
 
-    def _get_queue_deadletter_receiver(self, queue_name, **kwargs):
-        entity_name = generate_dead_letter_entity_name(
-            queue_name=queue_name,
-            is_transfer_deadletter=kwargs.get("transfer_deadletter", False)
-        )
-        return ServiceBusReceiver(
-            fully_qualified_namespace=self.fully_qualified_namespace,
-            entity_name=entity_name,
-            credential=self._credential,
-            logging_enable=self._config.logging_enable,
-            transport_type=self._config.transport_type,
-            http_proxy=self._config.http_proxy,
-            connection=self._connection,
-            **kwargs
-        )
-
-    def _get_subscription_deadletter_receiver(self, topic_name, subscription_name, **kwargs):
-        entity_name = generate_dead_letter_entity_name(
-            topic_name=topic_name,
-            subscription_name=subscription_name,
-            is_transfer_deadletter=kwargs.get("transfer_deadletter", False)
-        )
-        return ServiceBusReceiver(
-            fully_qualified_namespace=self.fully_qualified_namespace,
-            entity_name=entity_name,
-            credential=self._credential,
-            logging_enable=self._config.logging_enable,
-            transport_type=self._config.transport_type,
-            http_proxy=self._config.http_proxy,
-            connection=self._connection,
-            **kwargs
-        )
-
     def close(self):
         # type: () -> None
         """
@@ -176,8 +143,6 @@ class ServiceBusClient(object):
         :keyword int retry_total: The total number of attempts to redo a failed operation when an error occurs.
          Default value is 3.
         :rtype: ~azure.servicebus.ServiceBusSender
-        :raises: :class:`ServiceBusConnectionError`
-         :class:`ServiceBusAuthorizationError`
 
         .. admonition:: Example:
 
@@ -221,8 +186,6 @@ class ServiceBusClient(object):
         :keyword int retry_total: The total number of attempts to redo a failed operation when an error occurs.
          Default value is 3.
         :rtype: ~azure.servicebus.ServiceBusReceiver
-        :raises: :class:`ServiceBusConnectionError`
-         :class:`ServiceBusAuthorizationError`
 
         .. admonition:: Example:
 
@@ -247,10 +210,9 @@ class ServiceBusClient(object):
             **kwargs
         )
 
-
     def get_queue_deadletter_receiver(self, queue_name, **kwargs):
         # type: (str, Any) -> ServiceBusReceiver
-        """Get ServiceBusReceiver for the specific queue.
+        """Get ServiceBusReceiver for the sub Dead Letter Queue of the specific Queue.
 
         :param str queue_name: The path of specific Service Bus Queue the client connects to.
         :keyword mode: The mode with which messages will be retrieved from the entity. The two options
@@ -267,21 +229,21 @@ class ServiceBusClient(object):
          automatically shutdown. The default value is 0, meaning no timeout.
         :keyword int retry_total: The total number of attempts to redo a failed operation when an error occurs.
          Default value is 3.
-        :param transfer_deadletter: Whether to connect to the transfer deadletter queue, or the standard
+        :keyword float retry_backoff_factor: Delta back-off internal in the unit of second between retries.
+         Default value is 0.8.
+        :keyword float retry_backoff_max: Maximum back-off interval in the unit of second. Default value is 120.
+        :keyword bool transfer_deadletter: Whether to connect to the transfer deadletter queue, or the standard
          deadletter queue. Default is False, using the standard deadletter endpoint.
-        :type transfer_deadletter: bool
         :rtype: ~azure.servicebus.ServiceBusReceiver
-        :raises: :class:`ServiceBusConnectionError`
-         :class:`ServiceBusAuthorizationError`
 
         .. admonition:: Example:
 
             .. literalinclude:: ../samples/sync_samples/sample_code_servicebus.py
-                :start-after: [START create_servicebus_receiver_from_sb_client_sync]
-                :end-before: [END create_servicebus_receiver_from_sb_client_sync]
+                :start-after: [START create_queue_deadletter_receiver_from_sb_client_sync]
+                :end-before: [END create_queue_deadletter_receiver_from_sb_client_sync]
                 :language: python
                 :dedent: 4
-                :caption: Create a new instance of the ServiceBusReceiver from ServiceBusClient.
+                :caption: Create a new instance of the ServiceBusReceiver for Dead Letter Queue from ServiceBusClient.
 
 
         """
@@ -381,6 +343,57 @@ class ServiceBusClient(object):
             **kwargs
         )
 
+    def get_subscription_deadletter_receiver(self, topic_name, subscription_name, **kwargs):
+        """Get ServiceBusReceiver for the sub Dead Letter Queue of the specific subscription under the topic.
+
+        :param str topic_name: The name of specific Service Bus Topic the client connects to.
+        :param str subscription_name: The name of specific Service Bus Subscription
+         under the given Service Bus Topic.
+        :keyword mode: The mode with which messages will be retrieved from the entity. The two options
+         are PeekLock and ReceiveAndDelete. Messages received with PeekLock must be settled within a given
+         lock period before they will be removed from the subscription. Messages received with ReceiveAndDelete
+         will be immediately removed from the subscription, and cannot be subsequently rejected or re-received if
+         the client fails to process the message. The default mode is PeekLock.
+        :paramtype mode: ~azure.servicebus.ReceiveSettleMode
+        :keyword int prefetch: The maximum number of messages to cache with each request to the service.
+         The default value is 0, meaning messages will be received from the service and processed
+         one at a time. Increasing this value will improve message throughput performance but increase
+         the change that messages will expire while they are cached if they're not processed fast enough.
+        :keyword float idle_timeout: The timeout in seconds between received messages after which the receiver will
+         automatically shutdown. The default value is 0, meaning no timeout.
+        :keyword int retry_total: The total number of attempts to redo a failed operation when an error occurs.
+         Default value is 3.
+        :keyword float retry_backoff_factor: Delta back-off internal in the unit of second between retries.
+         Default value is 0.8.
+        :keyword float retry_backoff_max: Maximum back-off interval in the unit of second. Default value is 120.
+        :keyword bool transfer_deadletter: Whether to connect to the transfer deadletter queue, or the standard
+         deadletter queue. Default is False, using the standard deadletter endpoint.
+        :rtype: ~azure.servicebus.ServiceBusReceiver
+
+        .. admonition:: Example:
+
+            .. literalinclude:: ../samples/sync_samples/sample_code_servicebus.py
+                :start-after: [START create_subscription_deadletter_receiver_from_sb_client_sync]
+                :end-before: [END create_subscription_deadletter_receiver_from_sb_client_sync]
+                :language: python
+                :dedent: 4
+                :caption: Create a new instance of the ServiceBusReceiver for Dead Letter Queue from ServiceBusClient.
+
+
+        """
+        return ServiceBusReceiver(
+            fully_qualified_namespace=self.fully_qualified_namespace,
+            topic_name=topic_name,
+            subscription_name=subscription_name,
+            credential=self._credential,
+            logging_enable=self._config.logging_enable,
+            transport_type=self._config.transport_type,
+            http_proxy=self._config.http_proxy,
+            connection=self._connection,
+            is_dead_letter_receiver=True,
+            **kwargs
+        )
+
     def get_subscription_session_receiver(self, topic_name, subscription_name, session_id=None, **kwargs):
         # type: (str, str, str, Any) -> ServiceBusReceiver
         """Get ServiceBusReceiver for the specific subscription under the topic.
@@ -408,7 +421,7 @@ class ServiceBusClient(object):
         :keyword float retry_backoff_factor: Delta back-off internal in the unit of second between retries.
          Default value is 0.8.
         :keyword float retry_backoff_max: Maximum back-off interval in the unit of second. Default value is 120.
-        :rtype: ~azure.servicebus.ServiceBusReceiver
+        :rtype: ~azure.servicebus.ServiceBusSessionReceiver
 
         .. admonition:: Example:
 
@@ -460,8 +473,6 @@ class ServiceBusClient(object):
         :param int idle_timeout: The timeout in seconds between received messages after which the receiver will
          automatically shutdown. The default value is 0, meaning no timeout.
         :rtype: ~azure.servicebus.ServiceBusSessionReceiver
-        :raises: :class:`ServiceBusConnectionError`
-         :class:`ServiceBusAuthorizationError`
 
         .. admonition:: Example: