+++ sdk/servicebus/azure-servicebus/azure/servicebus/aio/_servicebus_receiver_async.py
@@ -247,6 +247,11 @@ class ServiceBusReceiver(collections.abc.AsyncIterator, BaseHandler, ReceiverMix
         :keyword dict http_proxy: HTTP proxy settings. This must be a dictionary with the following
          keys: `'proxy_hostname'` (str value) and `'proxy_port'` (int value).
          Additionally the following keys may also be present: `'username', 'password'`.
+        :keyword bool is_dead_letter_receiver: Should this receiver connect to the dead-letter-queue associated
+         with the specified entity, instead of the entity itself.  Default is `False`.
+        :keyword bool transfer_deadletter: Whether to connect to the transfer deadletter queue, or the standard
+         deadletter queue. Default is False, using the standard deadletter endpoint.  Only valid if
+         is_dead_letter_receiver is `True`, default is `False` regardless.
         :rtype: ~azure.servicebus.aio.ServiceBusReceiver
 
         .. admonition:: Example: