+++ sdk/servicebus/azure-servicebus/samples/async_samples/sample_code_servicebus_async.py
@@ -134,6 +134,16 @@ async def example_create_servicebus_receiver_async():
     )
     # [END create_servicebus_receiver_async]
 
+    # [START create_queue_deadletter_receiver_from_sb_client_async]
+    import os
+    from azure.servicebus.aio import ServiceBusClient
+    servicebus_connection_str = os.environ['SERVICE_BUS_CONNECTION_STR']
+    queue_name = os.environ['SERVICE_BUS_QUEUE_NAME']
+    servicebus_client = ServiceBusClient.from_connection_string(conn_str=servicebus_connection_str)
+    async with servicebus_client:
+        queue_receiver = servicebus_client.get_queue_deadletter_receiver(queue_name=queue_name)
+    # [END create_queue_deadletter_receiver_from_sb_client_async]
+
     # [START create_servicebus_receiver_from_sb_client_async]
     import os
     from azure.servicebus.aio import ServiceBusClient
@@ -143,6 +153,20 @@ async def example_create_servicebus_receiver_async():
     async with servicebus_client:
         queue_receiver = servicebus_client.get_queue_receiver(queue_name=queue_name)
     # [END create_servicebus_receiver_from_sb_client_async]
+
+    # [START create_subscription_deadletter_receiver_from_sb_client_async]
+    import os
+    from azure.servicebus import ServiceBusClient
+    servicebus_connection_str = os.environ['SERVICE_BUS_CONNECTION_STR']
+    topic_name = os.environ["SERVICE_BUS_TOPIC_NAME"]
+    subscription_name = os.environ["SERVICE_BUS_SUBSCRIPTION_NAME"]
+    servicebus_client = ServiceBusClient.from_connection_string(conn_str=servicebus_connection_str)
+    async with servicebus_client:
+        subscription_receiver = servicebus_client.get_subscription_deadletter_receiver(
+            topic_name=topic_name,
+            subscription_name=subscription_name,
+        )
+    # [END create_subscription_deadletter_receiver_from_sb_client_async]
 
     # [START create_subscription_receiver_from_sb_client_async]
     import os