+++ sdk/servicebus/azure-servicebus/tests/test_sessions.py
@@ -298,7 +298,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                     message.dead_letter(reason="Testing reason", description="Testing description")
 
             count = 0
-            with sb_client._get_queue_deadletter_receiver(servicebus_queue.name, idle_timeout=5) as receiver:
+            with sb_client.get_queue_deadletter_receiver(servicebus_queue.name, idle_timeout=5) as receiver:
                 for message in receiver:
                     count += 1
                     print_message(_logger, message)
@@ -408,7 +408,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                     messages = receiver.receive()
             assert count == 10
 
-            with sb_client._get_queue_deadletter_receiver(servicebus_queue.name,
+            with sb_client.get_queue_deadletter_receiver(servicebus_queue.name,
                                                       idle_timeout=5) as session:
                 count = 0
                 for message in session: