+++ sdk/servicebus/azure-servicebus/tests/test_subscriptions.py
@@ -114,7 +114,7 @@ class ServiceBusSubscriptionTests(AzureMgmtTestCase):
     @CachedServiceBusNamespacePreparer(name_prefix='servicebustest')
     @ServiceBusTopicPreparer(name_prefix='servicebustest')
     @ServiceBusSubscriptionPreparer(name_prefix='servicebustest')
-    def test_queue_by_servicebus_client_receive_batch_with_deadletter(self, servicebus_namespace_connection_string, servicebus_topic, servicebus_subscription, **kwargs):
+    def test_subscription_by_servicebus_client_receive_batch_with_deadletter(self, servicebus_namespace_connection_string, servicebus_topic, servicebus_subscription, **kwargs):
 
         with ServiceBusClient.from_connection_string(
                 servicebus_namespace_connection_string,
@@ -129,7 +129,7 @@ class ServiceBusSubscriptionTests(AzureMgmtTestCase):
                 prefetch=10
             ) as receiver:
 
-                with sb_client.get_queue_sender(servicebus_topic.name) as sender:
+                with sb_client.get_topic_sender(servicebus_topic.name) as sender:
                     for i in range(10):
                         message = Message("Dead lettered message no. {}".format(i))
                         sender.send(message)
@@ -158,7 +158,7 @@ class ServiceBusSubscriptionTests(AzureMgmtTestCase):
                     count += 1
             assert count == 0
 
-            with sb_client._get_subscription_deadletter_receiver(
+            with sb_client.get_subscription_deadletter_receiver(
                 topic_name=servicebus_topic.name,
                 subscription_name=servicebus_subscription.name,
                 idle_timeout=5,